borradas imagenes de prueba, cambiado ciertos elementos del responsive

+++ src/Controller/Gastronomia/ProductoTipicoDeleteController.php
@@ -19,13 +19,20 @@ class ProductoTipicoDeleteController extends AbstractController
     #[Route('/gastronomia/producto/delete/{uid}', name: 'app_producto_tipico_delete')]
     public function delete(ManagerRegistry $doctrine, ProductoTipico $producto): Response
     {
+
         $gastronomiaUid = $producto->getGastronomia()->getUid();
+        if ($producto == $producto->getGastronomia()->getProductoMes()) {
+            $this->addFlash("aviso", "No se puede borrar un producto seleccionado como 'Producto del mes'");
+            return $this->redirectToRoute('admin_gastronomia_get', [
+                'uid' => $gastronomiaUid
+            ]);
+        }
         $entityManager = $doctrine->getManager();
         $entityManager->remove($producto);
         $entityManager->flush();
         $this->addFlash("aviso", "Producto borrado correctamente");
         return $this->redirectToRoute('admin_gastronomia_get', [
-                'uid' => $gastronomiaUid
-            ]);
+            'uid' => $gastronomiaUid
+        ]);
     }
 }
